componentizise the htmls
