+++ beydamaxapp/app/producto/[slug]/components/Title.jsx
@@ -1,0 +1,7 @@
+export default function Title() {
+  return(
+    <div className="mt-4 border-b pb-6">
+      <h2 className="font-bold text-6xl">Milesstone Grill</h2>
+    </div>
+  )
+}
