+++ beydamaxapp/app/producto/[slug]/menu/page.jsx
@@ -1,66 +1,22 @@
+import NavBar from "@/app/components/NavBar";
 import Link from "next/link";
+import Header from "../components/Header";
+import ProductNavBar from "../components/ProductNavBar";
+import Menu from "../components/Menu";
 
 export default function MenuProducto() {
   return(
     <main className="bg-gray-100 min-h-screen w-screen">
       <main className="max-w-screen-2xl m-auto bg-white">
         {/* NAVBAR */}
-        <nav className="bg-white p-2 flex justify-between">
-          <Link href="/" className="font-bold text-gray-700 text-2xl">
-            Beydamax{" "}
-          </Link>
-          <div>
-            <div className="flex">
-              <button
-                className="bg-blue-400 text-white border p-1 px-4 rounded mr-3"
-              >
-                Sign in
-              </button>
-              <button className="border p-1 px-4 rounded">Sign up</button>
-            </div>
-          </div>
-        </nav>
-        {/* NAVBAR */} {/* HEADER */}
-        <div className="h-96 overflow-hidden">
-          <div
-            className="bg-center bg-gradient-to-r from-[#0f1f47] to-[#5f6984] h-full flex justify-center items-center"
-          >
-            <h1 className="text-7xl text-white captitalize text-shadow text-center">
-              Milestones Grill (Toronto)
-            </h1>
-          </div>
-        </div>
-        {/* HEADER */} {/* DESCRIPTION PORTION */}
+        < NavBar />
+        < Header />
         <div className="flex m-auto w-2/3 justify-between items-start 0 -mt-11">
           <div className="bg-white w-[100%] rounded p-3 shadow">
-            {/* RESAURANT NAVBAR */}
-            <nav className="flex text-reg border-b pb-2">
-              <Link href="/producto/milesstone-grill" className="mr-7"> Overview </Link>
-              <Link href="/producto/milesstone-grill/menu" className="mr-7"> Menu </Link>
-            </nav>
-            {/* RESAURANT NAVBAR */} {/* MENU */}
-            <main className="bg-white mt-5">
-              <div>
-                <div className="mt-4 pb-1 mb-1">
-                  <h1 className="font-bold text-4xl">Menu</h1>
-                </div>
-                <div className="flex flex-wrap justify-between">
-                  {/* MENU CARD */}
-                  <div className=" border rounded p-3 w-[49%] mb-3">
-                    <h3 className="font-bold text-lg">Surf and Turf</h3>
-                    <p className="font-light mt-1 text-sm">
-                      A well done steak with lobster and rice
-                    </p>
-                    <p className="mt-7">$80.00</p>
-                  </div>
-                  {/* MENU CARD */}
-                </div>
-              </div>
-            </main>
-            {/* MENU */}
+            < ProductNavBar />
+            < Menu />
           </div>
         </div>
-        {/* DESCRIPTION PORTION */}
       </main>
     </main>
   )
